# Slightly optimise nubOrd and add benchmark