--- a/benchmark/cabal.project
+++ b/benchmark/cabal.project
@@ -1,0 +1,1 @@
+packages: ../extra.cabal extra-benchmarks.cabal