I added a slight optimisation to nubOrd (based on the suggestion in Okasaki ex. 3.10a). I also added a benchmark (though I can't claim the inputs are very good or representative). Here are the numbers I got on my machine after this change:

Before:
```
benchmarking nubOrd/Int: Many repeats
time                 9.836 ms   (9.440 ms .. 10.20 ms)
                     0.993 R²   (0.988 R² .. 0.996 R²)
mean                 9.303 ms   (9.147 ms .. 9.481 ms)
std dev              455.6 μs   (366.9 μs .. 565.9 μs)
variance introduced by outliers: 22% (moderately inflated)

benchmarking nubOrd/Int: Few repeats 
time                 61.81 ms   (60.09 ms .. 63.63 ms)
                     0.998 R²   (0.995 R² .. 1.000 R²)
mean                 62.06 ms   (61.30 ms .. 63.23 ms)
std dev              1.804 ms   (1.195 ms .. 2.366 ms)

benchmarking nubOrd/String: Many repeats
time                 28.57 ms   (25.79 ms .. 30.84 ms)
                     0.983 R²   (0.974 R² .. 1.000 R²)
mean                 26.29 ms   (25.78 ms .. 27.32 ms)
std dev              1.503 ms   (506.2 μs .. 2.205 ms)
variance introduced by outliers: 21% (moderately inflated)

benchmarking nubOrd/String: Few repeats
time                 13.73 ms   (13.33 ms .. 14.03 ms)
                     0.996 R²   (0.991 R² .. 0.998 R²)
mean                 14.96 ms   (14.60 ms .. 15.43 ms)
std dev              991.7 μs   (800.9 μs .. 1.269 ms)
variance introduced by outliers: 31% (moderately inflated)
```

After:
```
benchmarking nubOrd/Int: Many repeats
time                 8.943 ms   (8.896 ms .. 8.986 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 8.993 ms   (8.968 ms .. 9.024 ms)
std dev              80.14 μs   (64.26 μs .. 98.31 μs)

benchmarking nubOrd/Int: Few repeats 
time                 57.67 ms   (56.56 ms .. 58.47 ms)
                     0.998 R²   (0.993 R² .. 1.000 R²)
mean                 57.97 ms   (57.30 ms .. 59.17 ms)
std dev              1.734 ms   (771.2 μs .. 2.785 ms)

benchmarking nubOrd/String: Many repeats
time                 25.58 ms   (25.00 ms .. 26.07 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 26.36 ms   (25.99 ms .. 26.91 ms)
std dev              1.012 ms   (775.6 μs .. 1.304 ms)

benchmarking nubOrd/String: Few repeats
time                 13.85 ms   (13.70 ms .. 13.96 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 13.91 ms   (13.83 ms .. 14.02 ms)
std dev              232.7 μs   (173.9 μs .. 343.7 μs)
```

More than happy to make any changes you think appropriate.