@ndmitchell 2020-06-09
Thanks! A nice little speed up, and thanks for all the benchmarking infrastructure.